+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [0.3.3-dev] - UNRELEASED
+### Added
+- Message cache for storing inbound messages
+- Link last inbound message to outbound messages for USSD flows
+- Retry messages if Turn rate limits us
 
 ## [0.3.2] - 2024-01-22
 ### Added